# Conflicts:
#	SERVER_MAFIA_HJ.py

+++ SERVER_MAFIA.py
@@ -156,10 +156,10 @@ class Job:
                     else:
                         vote_flag = True
                         sel_player = self.room.p_list[x]
-                        sendm(self.player, "{}(을)를 선택하셨습니다.".format(name_dic[sel_player]))
+                        sendm(self.player, " - {}(을)를 선택하셨습니다.".format(name_dic[sel_player]))
                         if self.name == '정치인':
                             sendm(self.player, "당신의 권력으로 두 표를 넣습니다.", line=False)
-                        broadcast(self.room.p_list, "{}에 한 표!".format(name_dic[sel_player]),
+                        broadcast(self.room.p_list, " - {}에 한 표!".format(name_dic[sel_player]),
                                   talker=[self.player])
                         if self.name == '정치인':
                             self.room.vote_list[x] += 1
@@ -220,31 +220,31 @@ class Job:
     @cerror_block
     def final_words(self):
         if self.player == self.room.vote_select:
-            broadcast(self.room.p_list, "{}의 최후의 한 마디가 있겠습니다!".format(name_dic[self.player]))
-            msg = recvm(self.player)
-            if self.room.timeout:
-                return
-            broadcast(self.room.p_list, "< {} : {} >".format(name_dic[self.player], msg), talker=[self.player],
-                      line=False)
-        else:
             while not self.room.timeout:
-                recvm(self.player)
+                msg = recvm(self.player)
                 if self.room.timeout:
                     return
+                broadcast(self.room.p_list, "< {} : {} >".format(name_dic[self.player], msg), talker=[self.player],
+                          line=False)
+        else:
+            if self.room.timeout:
+                return
 
     @cerror_block
     def final_vote(self):
         final_vote_flag = False
-        sendm(self.player, "{}(을)를 죽이는 데 찬성하시면 '찬성' 또는 'y',"
-                           "반대하시면 '반대' 또는 'n'을 입력하세요."
-                           "입력하지 않으면 반대표로 투표됩니다.".format(name_dic[self.room.vote_select]))
         while not self.room.timeout:
             msg = recvm(self.player)
             if self.room.timeout:
+                if self.name == '정치인':
+                    self.room.downvote += 1
                 self.room.downvote += 1
                 return
             if not final_vote_flag:
                 if msg == '찬성' or msg == 'Y' or msg == 'y':
+                    if self.name == '정치인':
+                        self.room.upvote += 1
+                        self.room.downvote -= 1
                     self.room.upvote += 1
                     self.room.downvote -= 1
                     sendm(self.player, "찬성하셨습니다!")
@@ -254,7 +254,8 @@ class Job:
                     sendm(self.player, "반대하셨습니다!")
                     final_vote_flag = True
                     continue
-            self.room.talk(self.player, msg)
+            if msg[-17:] != 'fEEBgFFDASDL%%@FM':
+                self.room.talk(self.player, msg)
 
     @cerror_block
     def select(self, player):
@@ -277,7 +278,6 @@ class Mafia(Job):
 
     @cerror_block
     def night_talk(self, msg):
-        print(name_dic[self.player])
         broadcast(self.room.mafia_list, '[MAFIA]{} : {}'.format(name_dic[self.player], msg), talker=[self.player])
         broadcast(self.room.dead_list, '[MAFIA]{} : {}'.format(name_dic[self.player], msg), talker=[self.player])
 
@@ -612,6 +612,7 @@ class Shaman(Job):
         if not self.use_skill:
             self.use_skill = True
             sendm(self.player, "{0}을 성불했습니다. {0}의 직업은 {1}입니다.".format(name_dic[player], self.room.job[player].name))
+            broadcast([player], "성불당했습니다.")
             self.room.job[player].shut_up = True
             return [True, player]
         else:
@@ -842,7 +843,9 @@ class Room:  # room 바로가기
             broadcast(self.p_list, "{}(이)가 투표로 죽었습니다.".format(name_dic[player]), talker=[player])
         if self.job[player].name == '테러리스트':
             if self.job[player].sel is not None:
-                self.kill(self.job[player].sel[1], 'by terrorist')
+                selected_player = self.job[player].sel[1]
+                if self.job[selected_player].name == '마피아':
+                    self.kill(self.job[player].sel[1], 'by terrorist')
             self.job[player].sel = None
         if self.job[player].name == '기자':
             self.news = None
@@ -864,7 +867,7 @@ class Room:  # room 바로가기
     def news_print(self):
         if self.news is not None:
             broadcast(self.p_list,
-                      "!!!!속보에요 속보!!!!\n{}(이)가 {}래요!".format(name_dic[self.news], self.job[self.news].name) + "#" * 100,
+                      "!!!!속보에요 속보!!!!\n{}(이)가 {}래요!\n".format(name_dic[self.news], self.job[self.news].name) + "#" * 100,
                       line_chr='#')
             self.news = None
 
@@ -886,15 +889,21 @@ class Room:  # room 바로가기
         else:
             broadcast(self.p_list, "{}번째 낮".format(day_num))
             self.news_print()
-            self.happening('morning', 10)
+            self.happening('morning', 15 * (self.player_num - len(self.dead_list)))
             broadcast(self.p_list, '투표 시간입니다! 투표할 사람을 선택해주세요. \n'
                                    '선택하는 방법은 "!(선택 번호)"를 입력해주시면 됩니다.')
-            self.happening('vote', 20)
+            self.happening('vote', 15)
             self.vote_result()
             voted_player = self.vote_select
             if voted_player is not None:
-                self.happening('final_words', 10)
-                self.happening('final_vote', 20)
+                broadcast(self.p_list, "{}의 최후의 한 마디가 있겠습니다!".format(name_dic[self.vote_select]))
+                self.happening('final_words', 15)
+                broadcast(self.p_list, "찬반 투표 시간입니다!\n"
+                                       "{}(을)를 죽이는 데 찬성하시면 '찬성' 또는 'y',\n"
+                                       "반대하시면 '반대' 또는 'n'을 입력하세요.\n"
+                                       "입력하지 않으면 반대표로 투표됩니다.".format(name_dic[self.vote_select])
+                          )
+                self.happening('final_vote', 10)
             if self.final_vote_result():
                 self.kill(voted_player, 'by vote')
             self.upvote, self.downvote, self.vote_list = 0, 0, [0] * self.player_num
@@ -951,8 +960,10 @@ class Room:  # room 바로가기
     def job_select(self):
         try:
             job_name_list = [Shaman, Terrorist, Soldier, Sherlock, Reporter, Politician]
-            job_num_dic = {Mafia: mafia_num[self.player_num], Terrorist: 2}
-            cnt = job_num_dic[Mafia] + 2
+            # job_num_dic = {Mafia: mafia_num[self.player_num], Reporter: 1, Terrorist: 1}
+            # cnt = job_num_dic[Mafia] + 2
+            job_num_dic = {Mafia: mafia_num[self.player_num], Shaman: 1}
+            cnt = job_num_dic[Mafia] + 1
             random.shuffle(job_name_list)
             for job_class in job_name_list:
                 if cnt >= self.player_num:
@@ -976,8 +987,6 @@ class Room:  # room 바로가기
                 job_index += 1
             for player in self.job:
                 print('{}:{}'.format(name_dic[player], self.job[player].name))
-            for player in self.mafia_list:
-                print(name_dic[player])
             return True
         except Exception as e:
             print(e)
